refactor(language): break up dialect.ts god-namespace
dialect.ts had grown to ~1700 lines, with several chunks that did not
actually need to live inside the Dialect class (which holds no state —
no fields, no constructor). This commit extracts four cohesive slices:

  core/discriminant.ts — DiscriminantConfig + prescanDiscriminantValue.
    Used by Block/NamedBlock factories via import, now imported from the
    new home. Re-exported from dialect.ts for back-compat.

  core/cst-diagnostics.ts — collectAllCstDiagnostics + missingNodeRange
    + collectCstDiagnosticsInner. Free functions that walk the parser
    CST collecting ERROR/MISSING diagnostics.

  core/comment-attacher.ts — parseInlineComments, parseElementComments,
    splitContainerComments, attachToFirstTypedMapEntry,
    attachToFirstProcedureStatement, attachToLastProcedureStatement,
    plus the hasProcedureStatements type guard. Previously private
    methods on Dialect that only used `this` to call a one-line
    parseComment delegator — now call sharedParseCommentNode directly.

Also removed the dead `Dialect.emit` method (28 lines, zero callers
across the monorepo — confirmed by grep of packages/, dialect/, apps/).
emitDocument in core/emit.ts is the real entry point.

No behavior change. Typecheck + full test suite + build pass.
